+++ src/Maybe/unwrapOr.ts
@@ -1,0 +1,5 @@
+import { Maybe, isNotNullAndUndefined } from './Maybe';
+
+export function unwrapOrFromMaybe<T>(v: Maybe<T>, def: T): T {
+    return isNotNullAndUndefined(v) ? v : def;
+}